refactored code and adding error handling format

+++ api/handlers/factura.go
@@ -11,20 +11,10 @@ import (
 	"os"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/jung-kurt/gofpdf"
 )
-
-func unmarshalServicios(data []byte) []models.Servicio {
-	var servicios []models.Servicio
-	err := json.Unmarshal(data, &servicios)
-	if err != nil {
-		fmt.Println("error al decodificar los datos JSON!", err)
-	}
-	return servicios
-}
 
 func ListarFacturas(c *gin.Context, db *storage.DB) {
 	// Obtener el rol del usuario del token JWT
@@ -33,99 +23,102 @@ func ListarFacturas(c *gin.Context, db *storage.DB) {
 
 	// Verificar si el usuario tiene el rol necesario para acceder a la ruta
 	if rol != "administrador" && rol != "usuario" {
-		c.JSON(http.StatusForbidden, gin.H{"error": "El usuario no tiene el rol necesario para acceder a esta ruta"})
+		errorResponse := models.ErrorResponseInit("INSUFFICIENT_ROLE", "Lo siento, pero parece que no tienes los permisos necesarios para acceder a esta página. Por favor, verifica tus credenciales o contacta al administrador para obtener más información.")
+		c.AbortWithStatusJSON(http.StatusForbidden, errorResponse)
 		return
 	}
+
+	// Obtener y validar los parámetros de la consulta
 	page, err := strconv.Atoi(c.DefaultQuery("page", "1"))
 	if err != nil || page < 1 {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "El parámetro 'page' debe ser un número entero positivo"})
+		errorResponse := models.ErrorResponseInit("INVALID_PAGE_PARAM", "El parámetro 'page' debe ser un número entero positivo")
+		c.AbortWithStatusJSON(http.StatusBadRequest, errorResponse)
 		return
 	}
 	limit, err := strconv.Atoi(c.DefaultQuery("limit", "10"))
 	if err != nil || limit < 1 {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "El parámetro 'limit' debe ser un número entero positivo"})
+		errorResponse := models.ErrorResponseInit("INVALID_LIMIT_PARAM", "El parámetro 'limit' debe ser un número entero positivo")
+		c.AbortWithStatusJSON(http.StatusBadRequest, errorResponse)
 		return
 	}
 	if limit > 100 {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "El parámetro 'limit' no puede ser mayor a 100"})
+		errorResponse := models.ErrorResponseInit("LIMIT_TOO_HIGH", "El parámetro 'limit' no puede ser mayor a 100")
+		c.AbortWithStatusJSON(http.StatusBadRequest, errorResponse)
 		return
 	}
+
+	// Calcular el desplazamiento y los filtros de la consulta
 	offset := (page - 1) * limit
 	if offset < 0 {
 		offset = 0
 	}
 	filterField := c.Query("filter_field")
 	filterValue := c.Query("filter_value")
+
+	// Construir y ejecutar la consulta para obtener las facturas
 	var rows *sql.Rows
+	var query string
+	var args []interface{}
 	if filterField != "" && filterValue != "" {
-		query := `SELECT * FROM facturas WHERE $1 = $2 ORDER BY id ASC LIMIT $3 OFFSET $4`
-		rows, err = db.Query(query, filterField, filterValue, limit, offset)
+		query = `SELECT * FROM facturas WHERE $1 = $2 ORDER BY id ASC LIMIT $3 OFFSET $4`
+		args = []interface{}{filterField, filterValue, limit, offset}
 	} else {
-		// Verificar si el usuario tiene el rol de "administrador"
 		if rol == "administrador" {
-			// Si el usuario es un administrador, listar todas las facturas
-			query := `SELECT * FROM facturas ORDER BY id ASC LIMIT $1 OFFSET $2`
-			rows, err = db.Query(query, limit, offset)
+			query = `SELECT * FROM facturas ORDER BY id ASC LIMIT $1 OFFSET $2`
+			args = []interface{}{limit, offset}
 		} else {
-			// Si el usuario no es un administrador, listar solo sus propias facturas
-			query := `SELECT * FROM facturas WHERE usuario_id = $1 ORDER BY id ASC LIMIT $2 OFFSET $3`
-			rows, err = db.Query(query, claims.UsuarioID, limit, offset)
+			query = `SELECT * FROM facturas WHERE usuario_id = $1 ORDER BY id ASC LIMIT $2 OFFSET $3`
+			args = []interface{}{claims.UsuarioID, limit, offset}
 		}
 	}
+	rows, err = db.Query(query, args...)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		errorResponse := models.ErrorResponseInit("DB_ERROR", "Error al obtener las facturas")
+		c.AbortWithStatusJSON(http.StatusInternalServerError, errorResponse)
+		c.Abort()
 		return
 	}
 	defer rows.Close()
-	var facturas []models.Factura
+
+	// Procesar las filas y construir el arreglo de facturas
+	facturas := make([]models.Factura, 0, limit)
 	for rows.Next() {
-		var (
-			id, usuarioID                     int
-			nombreEmpresa, nitEmpresa         string
-			fecha                             time.Time
-			servicios                         []byte
-			valorTotal                        float64
-			nombreOperador, tipoDocumento     string
-			documento, ciudadExpedicion       string
-			celular, numeroCuenta, tipoCuenta string
-			banco                             string
-		)
-		err := rows.Scan(&id, &nombreEmpresa, &nitEmpresa, &fecha, &servicios, &valorTotal, &nombreOperador, &tipoDocumento, &documento, &ciudadExpedicion, &celular, &numeroCuenta, &tipoCuenta, &banco, &usuarioID)
+		var factura models.Factura
+		err := factura.Scan(rows)
 		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			errorResponse := models.ErrorResponseInit("DB_ERROR", "Error al procesar las facturas")
+			c.AbortWithStatusJSON(http.StatusInternalServerError, errorResponse)
+			c.Abort()
 			return
-		}
-		factura := models.Factura{
-			ID:         id,
-			Empresa:    models.Empresa{Nombre: nombreEmpresa, NIT: nitEmpresa},
-			Fecha:      fecha,
-			Servicios:  unmarshalServicios(servicios),
-			ValorTotal: valorTotal,
-			Operador: models.Operador{Nombre: nombreOperador, TipoDocumento: tipoDocumento,
-				Documento: documento, CiudadExpedicionDocumento: ciudadExpedicion,
-				Celular:              celular,
-				NumeroCuentaBancaria: numeroCuenta,
-				TipoCuentaBancaria:   tipoCuenta, Banco: banco},
-			UsuarioID: int64(usuarioID),
 		}
 		facturas = append(facturas, factura)
 	}
+
+	// Verificar si se encontraron facturas y contar el total de facturas
 	switch len(facturas) {
 	case 0:
-		c.JSON(http.StatusNotFound, gin.H{"error": "La página solicitada no existe"})
+		errorResponse := models.ErrorResponseInit("NOT_FOUND", "La página solicitada no existe")
+		c.AbortWithStatusJSON(http.StatusNotFound, errorResponse)
+		c.Abort()
 	default:
 		var totalFacturas int
-		// Verificar si el usuario tiene el rol de "administrador"
-		if rol == "administrador" {
-			// Si el usuario es un administrador contar todas las facturas
-			err = db.QueryRow(`SELECT COUNT(*) FROM facturas`).Scan(&totalFacturas)
+		if filterField != "" && filterValue != "" {
+			query = `SELECT COUNT(*) FROM facturas WHERE $1 = $2`
+			args = []interface{}{filterField, filterValue}
 		} else {
-			// Si el usuario no es un administrador contar solo sus propias facturas
-			query := `SELECT COUNT(*) FROM facturas WHERE usuario_id = $1`
-			err = db.QueryRow(query, claims.UsuarioID).Scan(&totalFacturas)
+			if rol == "administrador" {
+				query = `SELECT COUNT(*) FROM facturas`
+				args = []interface{}{}
+			} else {
+				query = `SELECT COUNT(*) FROM facturas WHERE usuario_id = $1`
+				args = []interface{}{claims.UsuarioID}
+			}
 		}
+		err = db.QueryRow(query, args...).Scan(&totalFacturas)
 		if err != nil {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			errorResponse := models.ErrorResponseInit("DB_ERROR", "Error al contar las facturas")
+			c.AbortWithStatusJSON(http.StatusInternalServerError, errorResponse)
+			c.Abort()
 			return
 		}
 		totalPages := int(math.Ceil(float64(totalFacturas) / float64(limit)))
@@ -142,47 +135,134 @@ func CrearFactura(c *gin.Context, db *storage.DB) {
 	claims := c.MustGet("claims").(*models.Claims)
 	rol := claims.Role
 
-	// Verificar si el usuario tiene el rol necesario para acceder a la ruta
-	if rol != "administrador" && rol != "usuario" {
-		c.JSON(http.StatusForbidden, gin.H{"error": "El usuario no tiene el rol necesario para acceder a esta ruta"})
+	rolesPermitidos := []string{"administrador", "usuario"}
+	if !verificarRol(rol, rolesPermitidos) {
+		errorResponse := models.ErrorResponseInit("NO_PERMISSION", "No tienes permiso para acceder a esta página.")
+		c.JSON(http.StatusForbidden, errorResponse)
 		return
 	}
+
 	var factura models.Factura
-	err := c.BindJSON(&factura)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+	if err := c.BindJSON(&factura); err != nil {
+		errorResponse := models.ErrorResponseInit("INVALID_DATA", "Datos inválidos. Verifica y vuelve a intentarlo.")
+		c.JSON(http.StatusBadRequest, errorResponse)
 		return
 	}
 
 	serviciosJSON, err := json.Marshal(factura.Servicios)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		errorResponse := models.ErrorResponseInit("SERVICES_MARSHAL_ERROR", "Error al codificar los servicios en formato JSON.")
+		c.JSON(http.StatusInternalServerError, errorResponse)
 		return
 	}
 
 	query := `INSERT INTO facturas (nombre_empresa, nit_empresa, fecha, servicios, valor_total, nombre_operador, tipo_documento_operador, documento_operador, ciudad_expedicion_documento_operador, celular_operador, numero_cuenta_bancaria_operador, tipo_cuenta_bancaria_operador, banco_operador, usuario_id) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13,$14) RETURNING id`
-	err = db.QueryRow(query, factura.Empresa.Nombre, factura.Empresa.NIT, factura.Fecha, serviciosJSON, factura.ValorTotal, factura.Operador.Nombre, factura.Operador.TipoDocumento, factura.Operador.Documento, factura.Operador.CiudadExpedicionDocumento, factura.Operador.Celular, factura.Operador.NumeroCuentaBancaria, factura.Operador.TipoCuentaBancaria, factura.Operador.Banco, factura.UsuarioID).Scan(&factura.ID)
+	err = db.QueryRow(query,
+		factura.Empresa.Nombre,
+		factura.Empresa.NIT,
+		factura.Fecha,
+		serviciosJSON,
+		factura.ValorTotal,
+		factura.Operador.Nombre,
+		factura.Operador.TipoDocumento,
+		factura.Operador.Documento,
+		factura.Operador.CiudadExpedicionDocumento,
+		factura.Operador.Celular,
+		factura.Operador.NumeroCuentaBancaria,
+		factura.Operador.TipoCuentaBancaria,
+		factura.Operador.Banco,
+		factura.UsuarioID).Scan(&factura.ID)
+
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		errorResponse := models.ErrorResponseInit("DB_ERROR", "Error al procesar las facturas")
+		c.AbortWithStatusJSON(http.StatusInternalServerError, errorResponse)
+		c.Abort()
 		return
 	}
 
 	c.JSON(http.StatusCreated, gin.H{"message": "Factura creada correctamente", "factura": factura})
 }
 
+// obtenerUsuarioIDFactura consulta la base de datos para obtener el ID del usuario asociado a la factura especificada
+func obtenerUsuarioIDFactura(db *storage.DB, idFactura string) (int64, error) {
+	var idUsuarioFactura int64
+	err := db.QueryRow("SELECT usuario_id FROM facturas WHERE id = $1", idFactura).Scan(&idUsuarioFactura)
+	if err != nil {
+		return 0, err
+	}
+	return idUsuarioFactura, nil
+}
+
 func ActualizarFactura(c *gin.Context, db *storage.DB) {
-	id := c.Param("id")
+	// Obtener el rol y el ID del usuario del token JWT
+	claims := c.MustGet("claims").(*models.Claims)
+	rol := claims.Role
+	idUsuario := claims.UsuarioID
+	idFactura := c.Param("id")
+
+	// Verificar si el usuario tiene el rol necesario para acceder a la ruta
+	if !verificarRol(rol, []string{"administrador", "usuario"}) {
+		errorResponse := models.ErrorResponseInit("NO_PERMISSION", "No tienes permiso para acceder a esta página.")
+		c.JSON(http.StatusForbidden, errorResponse)
+		c.Abort()
+		return
+	}
+
+	// Si el rol del usuario es "usuario", verificar si está intentando actualizar su propia factura
+	if rol == "usuario" {
+		idUsuarioFactura, err := obtenerUsuarioIDFactura(db, idFactura)
+		if err != nil {
+			errorResponse := models.ErrorResponseInit("DB_ERROR", "Error al obtener el ID del usuario de la factura.")
+			c.JSON(http.StatusInternalServerError, errorResponse)
+			c.Abort()
+			return
+		}
+		if idUsuario != idUsuarioFactura {
+			errorResponse := models.ErrorResponseInit("NO_PERMISSION", "El usuario solo puede actualizar sus propias facturas.")
+			c.JSON(http.StatusForbidden, errorResponse)
+			c.Abort()
+			return
+		}
+	}
 
 	var factura models.Factura
 	err := c.BindJSON(&factura)
 	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		errorResponse := models.ErrorResponseInit("INVALID_DATA", "Datos inválidos. Verifica y vuelve a intentarlo.")
+		c.JSON(http.StatusBadRequest, errorResponse)
+		c.Abort()
+		return
+	}
+
+	// Validar los datos de entrada
+	if factura.Empresa.Nombre == "" || factura.Empresa.NIT == "" || factura.Fecha.IsZero() || len(factura.Servicios) == 0 {
+		errorResponse := models.ErrorResponseInit("MISSING_FIELDS", "Faltan campos requeridos.")
+		c.JSON(http.StatusBadRequest, errorResponse)
+		c.Abort()
 		return
 	}
 
 	serviciosJSON, err := json.Marshal(factura.Servicios)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		errorResponse := models.ErrorResponseInit("SERVICIOS_MARSHAL_ERROR", "Error al codificar los servicios en formato JSON.")
+		c.JSON(http.StatusInternalServerError, errorResponse)
+		c.Abort()
+		return
+	}
+
+	// Verificar si el usuario existe
+	var existeUsuario bool
+	err = db.QueryRow("SELECT EXISTS(SELECT 1 FROM usuarios WHERE id = $1)", factura.UsuarioID).Scan(&existeUsuario)
+	if err != nil {
+		errorResponse := models.ErrorResponseInit("DB_ERROR", "Error al verificar si el usuario existe.")
+		c.JSON(http.StatusInternalServerError, errorResponse)
+		c.Abort()
+		return
+	}
+	if !existeUsuario {
+		errorResponse := models.ErrorResponseInit("USER_NOT_FOUND", "El usuario especificado no existe.")
+		c.JSON(http.StatusBadRequest, errorResponse)
+		c.Abort()
 		return
 	}
 
@@ -196,9 +276,11 @@ func ActualizarFactura(c *gin.Context, db *storage.DB) {
     numero_cuenta_bancaria_operador = $11,
     tipo_cuenta_bancaria_operador = $12,
     banco_operador = $13, usuario_id=$14 WHERE id = $15`
-	result, err := db.Exec(query, factura.Empresa.Nombre, factura.Empresa.NIT, factura.Fecha, serviciosJSON, factura.ValorTotal, factura.Operador.Nombre, factura.Operador.TipoDocumento, factura.Operador.Documento, factura.Operador.CiudadExpedicionDocumento, factura.Operador.Celular, factura.Operador.NumeroCuentaBancaria, factura.Operador.TipoCuentaBancaria, factura.Operador.Banco, factura.UsuarioID, id)
+	result, err := db.Exec(query, factura.Empresa.Nombre, factura.Empresa.NIT, factura.Fecha, serviciosJSON, factura.ValorTotal, factura.Operador.Nombre, factura.Operador.TipoDocumento, factura.Operador.Documento, factura.Operador.CiudadExpedicionDocumento, factura.Operador.Celular, factura.Operador.NumeroCuentaBancaria, factura.Operador.TipoCuentaBancaria, factura.Operador.Banco, factura.UsuarioID, idFactura)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		errorResponse := models.ErrorResponseInit("DB_ERROR", "Error al actualizar la factura en la base de datos.")
+		c.JSON(http.StatusInternalServerError, errorResponse)
+		c.Abort()
 		return
 	}
 
@@ -210,23 +292,54 @@ func ActualizarFactura(c *gin.Context, db *storage.DB) {
 }
 
 func EliminarFactura(c *gin.Context, db *storage.DB) {
-	id := c.Param("id")
+	// Obtener el rol del usuario del token JWT
+	claims := c.MustGet("claims").(*models.Claims)
+	rol := claims.Role
+
+	// Verificar si el usuario tiene el rol necesario para acceder a la ruta
+	rolesPermitidos := []string{"administrador", "usuario"}
+	if !verificarRol(rol, rolesPermitidos) {
+		errorResponse := models.ErrorResponseInit("NO_PERMISSION", "No tienes permiso para acceder a esta página.")
+		c.JSON(http.StatusForbidden, errorResponse)
+		c.Abort()
+		return
+	}
+
+	// Validar el valor del parámetro id
+	idStr := c.Param("id")
+	id, err := strconv.Atoi(idStr)
+	if err != nil {
+		errorResponse := models.ErrorResponseInit("INVALID_ID", "El valor del parámetro id no es un número válido")
+		c.JSON(http.StatusBadRequest, errorResponse)
+		c.Abort()
+		return
+	}
+	if id <= 0 {
+		errorResponse := models.ErrorResponseInit("INVALID_ID", "El valor del parámetro id debe ser un número entero positivo")
+		c.JSON(http.StatusBadRequest, errorResponse)
+		c.Abort()
+		return
+	}
 
 	query := `DELETE FROM facturas WHERE id = $1`
 	result, err := db.Exec(query, id)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		errorResponse := models.ErrorResponseInit("SQL_ERROR", "Error al ejecutar la consulta SQL")
+		c.JSON(http.StatusInternalServerError, errorResponse)
+		c.Abort()
 		return
 	}
 
 	if rowsAffected, _ := result.RowsAffected(); rowsAffected > 0 {
 		c.JSON(http.StatusOK, gin.H{"message": "Factura eliminada correctamente"})
 	} else {
-		c.JSON(http.StatusNotFound, gin.H{"error": "No se encontró la factura con el ID especificado"})
+		errorResponse := models.ErrorResponseInit("NOT_FOUND", "No se encontró la factura con el ID especificado")
+		c.JSON(http.StatusNotFound, errorResponse)
+		c.Abort()
 	}
 }
 
-func GenerarPDF(c *gin.Context, db *storage.DB) {
+func obtenerFactura(c *gin.Context, db *storage.DB) (models.Factura, error) {
 	// Obtener el ID de la factura a partir del parámetro de la URL
 	id := c.Param("id")
 
@@ -242,19 +355,48 @@ func GenerarPDF(c *gin.Context, db *storage.DB) {
 	if err != nil {
 		if err == sql.ErrNoRows {
 			// Manejar el caso en que no hay filas para escanear
-			c.JSON(http.StatusNotFound, gin.H{"error": "Factura no encontrada"})
-			return
+			return factura, fmt.Errorf("factura no encontrada")
 		} else {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-			return
+			return factura, err
 		}
 	}
 
 	// Decodificar los datos JSON de los servicios en una slice de estructuras Servicio
 	err = json.Unmarshal(serviciosJSON, &factura.Servicios)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return factura, err
+	}
+
+	return factura, nil
+}
+
+func GenerarPDF(c *gin.Context, db *storage.DB) {
+	// Obtener el ID de la factura a partir del parámetro de la URL
+	id := c.Param("id")
+
+	// Obtener el rol del usuario del token JWT
+	claims := c.MustGet("claims").(*models.Claims)
+	rol := claims.Role
+
+	// Verificar si el usuario tiene el rol necesario para acceder a la ruta
+	rolesPermitidos := []string{"administrador", "usuario"}
+	if !verificarRol(rol, rolesPermitidos) {
+		errorResponse := models.ErrorResponseInit("NO_PERMISSION", "No tienes permiso para acceder a esta página.")
+		c.JSON(http.StatusForbidden, errorResponse)
+		c.Abort()
 		return
+	}
+
+	// Obtener la factura
+	factura, err := obtenerFactura(c, db)
+	if err != nil {
+		if err.Error() == "Factura no encontrada" {
+			c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
+			return
+		} else {
+			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+			return
+		}
 	}
 
 	// Crear un nuevo documento PDF
@@ -360,4 +502,13 @@ func GenerarPDF(c *gin.Context, db *storage.DB) {
 
 	// Enviar el archivo PDF como respuesta
 	c.File(tmpfile.Name())
+}
+
+func verificarRol(rol string, rolesPermitidos []string) bool {
+	for _, rolPermitido := range rolesPermitidos {
+		if rol == rolPermitido {
+			return true
+		}
+	}
+	return false
 }
